Add empty CargoDoorFwd object to CRJ700 model
This avoids the warning message on CRJ700 startup about the
missing CargoDoorFwd on the CRJ700.

Note: This does not fix the number of cargo doors on the EICAS
display in the CRJ700, though.

Context information:

The CRJ700 has two cargo doors (CargoDoorCtr and CargoDoorAft)
while the CRJ900* and CRJ1000* have three cargo doors (CargoDoorFwd,
CargoDoorCtr, and CargoDoorAft).

Also, the number of emergency exits over the wing differs between
the models in CRJ700 family. Those are not being modelled yet, though.